perf: 优化性能,修改文档与重构部分结构
