+++ src/service/mod.rs
@@ -1,5 +1,4 @@
 pub mod maimai_best_50;
 pub mod printer;
 pub mod resource;
-pub mod simple_log;
 pub mod table;
